refactor: separate bubbles to its module
